Przycisk zapisu stanu gry pod ikonę skarbów Resetowanie liczby skarbów
