+++ RumbleJungle/RumbleJungle/Skins/MainSkin.xaml
@@ -20,6 +20,12 @@
                                           VerticalAlignment="Center" />
                     </Border>
                     <ControlTemplate.Triggers>
+                        <Trigger Property="IsDefaulted"
+                                 Value="True">
+                            <Setter Property="BorderBrush"
+                                    TargetName="border"
+                                    Value="{DynamicResource {x:Static SystemColors.HighlightBrushKey}}" />
+                        </Trigger>
                         <Trigger Property="IsMouseOver"
                                  Value="True">
                             <Setter TargetName="border"
@@ -28,6 +34,35 @@
                             <Setter TargetName="border"
                                     Property="Background"
                                     Value="#FFCDEEF9" />
+                        </Trigger>
+                        <Trigger Property="IsPressed"
+                                 Value="True">
+                            <Setter Property="Background"
+                                    TargetName="border"
+                                    Value="#FFC4E5F6" />
+                            <Setter Property="BorderBrush"
+                                    TargetName="border"
+                                    Value="#FF2C628B" />
+                        </Trigger>
+                        <Trigger Property="ToggleButton.IsChecked"
+                                 Value="True">
+                            <Setter Property="Background"
+                                    TargetName="border"
+                                    Value="#FFBCDDEE" />
+                            <Setter Property="BorderBrush"
+                                    TargetName="border"
+                                    Value="#FF245A83" />
+                        </Trigger>
+                        <Trigger Property="IsEnabled"
+                                 Value="False">
+                            <Setter Property="Background"
+                                    TargetName="border"
+                                    Value="#FFF4F4F4" />
+                            <Setter Property="BorderBrush"
+                                    TargetName="border"
+                                    Value="#FFADB2B5" />
+                            <Setter Property="Foreground"
+                                    Value="#FF838383" />
                         </Trigger>
                     </ControlTemplate.Triggers>
                 </ControlTemplate>
