+++ RumbleJungle/RumbleJungle/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="CommonServiceLocator" version="2.0.4" targetFramework="net472" />
+  <package id="CommonServiceLocator" version="2.0.5" targetFramework="net472" />
   <package id="Microsoft.CodeAnalysis.FxCopAnalyzers" version="2.9.7" targetFramework="net472" developmentDependency="true" />
   <package id="Microsoft.CodeAnalysis.VersionCheckAnalyzer" version="2.9.7" targetFramework="net472" developmentDependency="true" />
   <package id="Microsoft.CodeQuality.Analyzers" version="2.9.7" targetFramework="net472" developmentDependency="true" />
